requirements.txt
aiogram==3.4.1
aiohttp==3.9
aiohttp_session==2.12
# multi-file-converter==1.0.9
python-dateutil==2.8.2
# python-docx==0.8.11
python-dotenv==1.0.0
pytz==2024.1
cryptography==42.0.5
requests==2.32.3